Update api for BulkWriteResult
Fixes JAVA-1191
