+++ driver-compat/src/main/com/mongodb/UnacknowledgedBulkWriteResult.java
@@ -34,7 +34,7 @@ class UnacknowledgedBulkWriteResult extends BulkWriteResult {
     }
 
     @Override
-    public int getUpdatedCount() {
+    public int getMatchedCount() {
         throw getUnacknowledgedWriteException();
     }
