+++ driver-compat/src/test/unit/com/mongodb/DBCollectionTest.java
@@ -605,7 +605,7 @@ public class DBCollectionTest extends DatabaseTestCase {
         assertTrue(bulkWriteOperation.isOrdered());
         assertTrue(result.isAcknowledged());
         assertEquals(1, result.getInsertedCount());
-        assertEquals(4, result.getUpdatedCount());
+        assertEquals(4, result.getMatchedCount());
         assertEquals(3, result.getRemovedCount());
         if (result.isModifiedCountAvailable()) {
             assertEquals(4, result.getModifiedCount());
