+++ driver/src/test/functional/org/mongodb/operation/MixedBulkWriteOperationSpecification.groovy
@@ -16,7 +16,7 @@
 
 
 package org.mongodb.operation
-import category.Async
+
 import category.Slow
 import org.bson.types.ObjectId
 import org.junit.experimental.categories.Category
@@ -33,24 +33,22 @@ import org.mongodb.codecs.PrimitiveCodecs
 import org.mongodb.protocol.AcknowledgedBulkWriteResult
 
 import static org.junit.Assume.assumeTrue
-import static org.mongodb.Fixture.getAsyncBinding
+import static org.mongodb.Fixture.getBinding
 import static org.mongodb.Fixture.serverVersionAtLeast
 import static org.mongodb.WriteConcern.ACKNOWLEDGED
 import static org.mongodb.WriteConcern.UNACKNOWLEDGED
 import static org.mongodb.operation.WriteRequest.Type.REMOVE
 import static org.mongodb.operation.WriteRequest.Type.UPDATE
 
+class MixedBulkWriteOperationSpecification extends FunctionalSpecification {
 
-@Category(Async)
-class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
-
-    def  'when no document with the same id exists, should insert the document'() {
+    def 'when no document with the same id exists, should insert the document'() {
         given:
         def op = new MixedBulkWriteOperation(getNamespace(), [new InsertRequest<Document>(new Document('_id', 1))], ordered,
                                              ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result.insertedCount == 1
@@ -61,7 +59,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when a document with the same id exists, should throw an exception'() {
+    def 'when a document with the same id exists, should throw an exception'() {
         given:
         def document = new Document('_id', 1)
         getCollectionHelper().insertDocuments(document)
@@ -69,7 +67,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         then:
         def ex = thrown(BulkWriteException)
@@ -79,14 +77,14 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when a document with no _id  is inserted and a Id generating codec is supplied the _id should be generated by the driver'() {
+    def 'when a document with no _id  is inserted and a Id generating codec is supplied the _id should be generated by the driver'() {
         given:
         def document = new Document()
         def encoder = new CollectibleDocumentCodec(PrimitiveCodecs.createDefault(), new ObjectIdGenerator())
         def op = new MixedBulkWriteOperation(getNamespace(), [new InsertRequest<Document>(document)], ordered, ACKNOWLEDGED, encoder)
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result.insertedCount == 1
@@ -98,7 +96,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when documents match the query, a remove of one should remove one of them'() {
+    def 'when documents match the query, a remove of one should remove one of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true))
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -106,7 +104,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(REMOVE, 1, [])
@@ -116,7 +114,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when documents match the query, a remove should remove all of them'() {
+    def 'when documents match the query, a remove should remove all of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true), new Document('x', false))
 
@@ -125,7 +123,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(REMOVE, 2, [])
@@ -135,7 +133,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when multiple document match the query, update of one should update only one of them'() {
+    def 'when multiple document match the query, update of one should update only one of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true));
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -144,7 +142,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 1, expectedModifiedCount(1), [])
@@ -154,7 +152,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when documents match the query, update multi should update all of them'() {
+    def 'when documents match the query, update multi should update all of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true));
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -163,7 +161,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 2, expectedModifiedCount(2), [])
@@ -173,7 +171,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when no document matches the query, an update of one with upsert should insert a document'() {
+    def 'when no document matches the query, an update of one with upsert should insert a document'() {
         def id = new ObjectId()
         def query = new Document('_id', id)
         given:
@@ -182,7 +180,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 0, expectedModifiedCount(0), [new BulkWriteUpsert(0, id)])
@@ -192,7 +190,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when no document matches the query, an update multi with upsert should insert a document'() {
+    def 'when no document matches the query, an update multi with upsert should insert a document'() {
         def id = new ObjectId()
         def query = new Document('_id', id)
         given:
@@ -202,7 +200,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 0, expectedModifiedCount(0), [new BulkWriteUpsert(0, id)])
@@ -212,7 +210,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when documents matches the query, update one with upsert should update only one of them'() {
+    def 'when documents matches the query, update one with upsert should update only one of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true));
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -221,7 +219,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 1, expectedModifiedCount(1), [])
@@ -231,7 +229,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when documents match the query, update multi with upsert should update all of them'() {
+    def 'when documents match the query, update multi with upsert should update all of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true));
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -240,7 +238,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 2, expectedModifiedCount(2), [])
@@ -250,7 +248,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when a document contains a key with an illegal character, replacing a document with it should throw IllegalArgumentException'() {
+    def 'when a document contains a key with an illegal character, replacing a document with it should throw IllegalArgumentException'() {
         given:
         def id = new ObjectId()
         def encoder = new CollectibleDocumentCodec(PrimitiveCodecs.createDefault(), new ObjectIdGenerator())
@@ -260,7 +258,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              true, ACKNOWLEDGED, encoder)
 
         when:
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         then:
         thrown(IllegalArgumentException)
@@ -269,7 +267,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when no document matches the query, a replace with upsert should insert a document'() {
+    def 'when no document matches the query, a replace with upsert should insert a document'() {
         given:
         def id = new ObjectId()
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -278,7 +276,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 0, expectedModifiedCount(0), [new BulkWriteUpsert(0, id)])
@@ -288,7 +286,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'when multiple documents match the query, replace should replace only one of them'() {
+    def 'when multiple documents match the query, replace should replace only one of them'() {
         given:
         getCollectionHelper().insertDocuments(new Document('x', true), new Document('x', true))
 
@@ -298,7 +296,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ordered, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 1, expectedModifiedCount(1), [])
@@ -308,8 +306,8 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    @Category([Async, Slow])
-    def  'when a replacement document is 16MB, the document is still replaced'() {
+    @Category(Slow)
+    def 'when a replacement document is 16MB, the document is still replaced'() {
         given:
         getCollectionHelper().insertDocuments(new Document('_id', 1))
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -319,15 +317,15 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              true, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 1, expectedModifiedCount(1), [])
         getCollectionHelper().count() == 1
     }
 
-    @Category([Async, Slow])
-    def  'when two update documents together exceed 16MB, the documents are still updated'() {
+    @Category(Slow)
+    def 'when two update documents together exceed 16MB, the documents are still updated'() {
         given:
         getCollectionHelper().insertDocuments(new Document('_id', 1), new Document('_id', 2))
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -338,14 +336,14 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              true, ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result == new AcknowledgedBulkWriteResult(UPDATE, 2, expectedModifiedCount(2), [])
         getCollectionHelper().count() == 2
     }
 
-    def  'should handle multi-length runs of ordered insert, update, replace, and remove'() {
+    def 'should handle multi-length runs of ordered insert, update, replace, and remove'() {
         given:
         getCollectionHelper().insertDocuments(getTestInserts())
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -353,7 +351,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         result.acknowledged
@@ -370,7 +368,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'should handle multi-length runs of unacknowledged insert, update, replace, and remove'() {
+    def 'should handle multi-length runs of unacknowledged insert, update, replace, and remove'() {
         given:
         getCollectionHelper().insertDocuments(getTestInserts())
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -378,7 +376,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              UNACKNOWLEDGED, new DocumentCodec())
 
         when:
-        def result = op.executeAsync(getAsyncBinding()).get()
+        def result = op.execute(getBinding())
 
         then:
         !result.acknowledged
@@ -396,7 +394,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'should split the number of writes is larger than the match write batch size'() {
+    def 'should split the number of writes is larger than the match write batch size'() {
         given:
         def writes = []
         (0..2000).each {
@@ -404,9 +402,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         }
 
         when:
-        new MixedBulkWriteOperation(getNamespace(), writes, ordered, ACKNOWLEDGED, new DocumentCodec())
-                .executeAsync(getAsyncBinding())
-                .get()
+        new MixedBulkWriteOperation(getNamespace(), writes, ordered, ACKNOWLEDGED, new DocumentCodec()).execute(getBinding())
 
         then:
         getCollectionHelper().count() == 2001
@@ -415,7 +411,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'error details should have correct index on ordered write failure'() {
+    def 'error details should have correct index on ordered write failure'() {
         given:
         def op = new MixedBulkWriteOperation(getNamespace(),
                                              [new InsertRequest<Document>(new Document('_id', 1)),
@@ -423,7 +419,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                               new InsertRequest<Document>(new Document('_id', 1))   // this should fail with index 2
                                              ], true, ACKNOWLEDGED, new DocumentCodec())
         when:
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         then:
         def ex = thrown(BulkWriteException)
@@ -432,7 +428,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ex.writeErrors[0].code == 11000
     }
 
-    def  'error details should have correct index on unordered write failure'() {
+    def 'error details should have correct index on unordered write failure'() {
         given:
         getCollectionHelper().insertDocuments(getTestInserts())
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -441,7 +437,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                               new InsertRequest<Document>(new Document('_id', 3))   // this should fail with index 2
                                              ], false, ACKNOWLEDGED, new DocumentCodec())
         when:
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         then:
         def ex = thrown(BulkWriteException)
@@ -453,7 +449,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
     }
 
     // using w = 5 to force a timeout
-    def  'should throw bulk write exception with a write concern error when wtimeout is exceeded'() {
+    def 'should throw bulk write exception with a write concern error when wtimeout is exceeded'() {
         assumeTrue(Fixture.isDiscoverableReplicaSet())
         given:
         def op = new MixedBulkWriteOperation(getNamespace(),
@@ -461,14 +457,14 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              false, new WriteConcern(5, 1), new DocumentCodec()
         )
         when:
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         then:
         def ex = thrown(BulkWriteException)
         ex.getWriteConcernError() != null
     }
 
-    def  'when there is a duplicate key error and a write concern error, both should be reported'() {
+    def 'when there is a duplicate key error and a write concern error, both should be reported'() {
         assumeTrue(Fixture.isDiscoverableReplicaSet())
 
         given:
@@ -479,7 +475,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
                                              ], ordered, new WriteConcern(4, 1), new DocumentCodec())
 
         when:
-        op.executeAsync(getAsyncBinding()).get()  // This is assuming that it won't be able to replicate to 4 servers in 1 ms
+        op.execute(getBinding())  // This is assuming that it won't be able to replicate to 4 servers in 1 ms
 
         then:
         def ex = thrown(BulkWriteException)
@@ -492,16 +488,16 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [false]
     }
 
-    def  'execute should throw IllegalStateException when already executed'() {
+    def 'execute should throw IllegalStateException when already executed'() {
         given:
         def op = new MixedBulkWriteOperation(getNamespace(),
                                              [new InsertRequest<Document>(new Document('_id', 1))],
                                              ordered, UNACKNOWLEDGED, new DocumentCodec())
 
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         when:
-        op.executeAsync(getAsyncBinding()).get()
+        op.execute(getBinding())
 
         then:
         thrown(IllegalStateException)
@@ -510,7 +506,7 @@ class BulkWriteOperationAsyncSpecification extends FunctionalSpecification {
         ordered << [true, false]
     }
 
-    def  'should throw IllegalArgumentException when passed an empty bulk operation'() {
+    def 'should throw IllegalArgumentException when passed an empty bulk operation'() {
 
         when:
         new MixedBulkWriteOperation(getNamespace(), [], ordered, UNACKNOWLEDGED, new DocumentCodec())
